Rename externalGraderEC2 -> externalGraderECS
