+++ lib/externalGrader.js
@@ -7,7 +7,7 @@ const logger = require('./logger');
 const sqldb = require('./sqldb');
 const sqlLoader = require('./sql-loader');
 const externalGradingSocket = require('./external-grading-socket');
-const ExternalGraderEC2 = require('./externalGraderEC2');
+const ExternalGraderECS = require('./externalGraderECS');
 const ExternalGraderBatch = require('./externalGraderBatch');
 const ExternalGraderLocal = require('./externalGraderLocal');
 
@@ -30,7 +30,7 @@ module.exports.init = function(assessment, callback) {
             module.exports.grader = new ExternalGraderBatch();
         } else {
             AWS.config.update({region: 'us-east-2'});
-            module.exports.grader = new ExternalGraderEC2();
+            module.exports.grader = new ExternalGraderECS();
         }
         callback(null);
     } else {
